Move RWUTable to a separate module
